+++ tardis/urls.py
@@ -271,11 +271,13 @@ group_urls = patterns(
 facility_urls = patterns(
     'tardis.tardis_portal.views',
     (r'^overview/$', 'facility_overview'),
-    (r'^fetch_data/(?P<facility_id>\d+)/count/', 'fetch_facility_data_count'),
+    (r'^fetch_data/(?P<facility_id>\d+)/count/', 'facility_overview_data_count'),
     (r'^fetch_data/(?P<facility_id>\d+)/'
      r'(?P<start_index>\d+)/(?P<end_index>\d+)/$',
-     'fetch_facility_data'),
-    (r'^fetch_facilities_list/$', 'fetch_facilities_list'),
+     'facility_overview_experiments'),
+    (r'^fetch_datafiles/(?P<dataset_id>\d+)/$',
+     'facility_overview_dataset_detail'),
+    (r'^fetch_facilities_list/$', 'facility_overview_facilities_list'),
     )
 
 display_urls = patterns(